Migrate slant_direction from GoogleFonts to OpenType profile.
opentype/slant_direction: "Checking direction of slnt axis angles."
