+++ Lib/fontbakery/profiles/microsoft.py
@@ -7,6 +7,7 @@ PROFILE = {
     ],
     "pending_review": [
         "opentype:cff_ascii_strings",
+        "opentype:slant_direction",
         "opentype:weight_class_fvar",
         "designspace_has_consistent_codepoints",
         "designspace_has_consistent_glyphset",
